Move generic options code to supporting class
Re #23051
