+++ qt/scientific_interfaces/ISISReflectometry/Common/CMakeLists.txt
@@ -1,5 +1,5 @@
 set(COMMON_SRC_FILES Parse.cpp GetInstrumentParameter.cpp
-                     InstrumentParameters.cpp)
+                     InstrumentParameters.cpp OptionDefaults.cpp)
 
 # Include files aren't required, but this makes them appear in Visual Studio
 # IMPORTANT: Include files are required in the MOC_FILES set. Scroll down to
@@ -12,6 +12,7 @@ set(COMMON_INC_FILES
     InstrumentParameters.h
     Map.h
     IMessageHandler.h
+    OptionDefaults.h
     Parse.h
     QWidgetGroup.h
     ValidationResult.h
